requirements.txt
anyio==3.7.0
beautifulsoup4==4.12.2
certifi==2023.5.7
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.99.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==1.10.10
python-telegram-bot==20.4
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0